transfer detail view for explorer and removing explorer code from UI
