--- packages/example/src/components/DetailView/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./DetailView";
