--- packages/example/src/components/ExplorerTable/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ExplorerTable";
