--- packages/example/src/pages/ExplorerPage/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ExplorerPage";
