+++ packages/example/src/pages/index.ts
@@ -1,4 +1,3 @@
-export { default as ExplorerPage } from "./ExplorerPage";
 export { default as TransactionPage } from "./TransactionPage";
 export { default as TransferPage } from "./TransferPage";
 export { default as WrapperPage } from "./WrapperPage";
