+++ packages/explorer-ui/src/components/TransferDetailView/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TransferDetailView";
