+++ packages/explorer-ui/src/components/index.ts
@@ -1,2 +1,3 @@
 export { default as ExplorerTable } from "./ExplorerTable";
 export { default as Header } from './Header'
+export { default as TransferDetailView } from './TransferDetailView'
