+++ packages/explorer-ui/src/pages/TransactionPage/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./TransactionPage";
