+++ packages/explorer-ui/src/pages/TransactionPage/styles.ts
@@ -1,0 +1,11 @@
+import { makeStyles, createStyles } from "@chainsafe/common-theme";
+
+export const useStyles = makeStyles(() =>
+  createStyles({
+    transferDetailViewContainer: {
+      display: "flex",
+      flexDirection: "row",
+      justifyContent: "center",
+    },
+  })
+);
