+++ packages/explorer-ui/src/pages/index.ts
@@ -1,1 +1,2 @@
 export { default as ExplorerPage } from "./ExplorerPage";
+export { default as TransactionPage } from "./TransactionPage";
